requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
limits==5.8.0
packaging==26.0
pluggy==1.6.0
pyaes==1.6.1
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
rsa==4.9.1
slowapi==0.1.9
starlette==0.52.1
structlog==25.5.0
Telethon==1.42.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
wrapt==2.1.1
